pyproject.toml
[tool.poetry]
name = "vilka-backend"
version = "0.1.0"
description = ""
authors = ["Oleg <mikus.oleg@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"
SQLAlchemy = "^1.4.40"
alembic = "^1.8.1"
python-dotenv = "^0.20.0"
psycopg2-binary = "^2.9.3"
python-multipart = "^0.0.5"
pyjwt = "^2.6.0"
passlib = "^1.7.4"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
aerich = "^0.7.1"
bcrypt = "^4.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
safety = "^2.1.1"
flake8 = "5.0.4"
mypy = "^0.971"
pre-commit = "^2.12.1"
autoflake = "^1.4"
autopep8 = "^1.7.0"
isort = "==5.10.1"
bandit = "1.7.4"
funcy = "^1.16"
flake8-bandit = "4.1.1"


[tool.aerich]
tortoise_orm = "src.config.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 160

[tool.mypy]
disallow_untyped_defs = true
follow_imports = 'silent'
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = false
warn_no_return = true
warn_unreachable = true
strict_equality = true
local_partial_types = true
ignore_errors = false
allow_redefinition = false
